fin_savvy_app/requirements.txt
fastapi==0.133.1
greenlet==3.3.2
psycopg2-binary==2.9.11
pydantic==2.12.5
SQLAlchemy==2.0.47
starlette==0.52.1
itsdangerous>=2.0,<3
typing_extensions==4.15.0
uvicorn==0.41.0
python-multipart==0.0.20
pdfplumber==0.11.4
Jinja2
bcrypt>=4.0
# Local ML classifier (train from CSV; no API)
scikit-learn>=1.3.0
joblib>=1.3.0
# Optional: for API-based classification later (FINSAVVY_CLASSIFIER=openai + OPENAI_API_KEY)
openai>=1.0.0
# Budget insights (category summaries, daily series)
pandas>=2.0.0
# Optional receipt OCR: also install system package `tesseract-ocr` (see docs/CORE_FEATURES.md)
Pillow>=10.0.0
pytesseract>=0.3.10
# Optional: generate docs/FINSAVVY_FEATURE_ROADMAP_*.docx — python scripts/build_feature_roadmap_docx.py
python-docx>=1.1.0